+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add new service to handle route53 DNS records.
 - Add new service to handle ACM certificates.
+- Use predictable domain alias for cloudfront on legacy clusters.
 
 ### Changed